Move timer fixture to conftest.py and support generic callbacks
